Ensure enqueued items never precede dequeued items
This change strengthens an invariant that was previously only upheld
within the current segment. This invariant ensures that when a segment
is written (or segments are merged) that all of the entries in the
segment which have been dequeued will be at the start of the segment.
Without this invariant it would be possible for the dequeued entries
within a segment to be discontinuous, making it difficult to skip over
them when re-opening a freshly written segment.