Reimplement ID generator to manage its own state
Replaces `StandardIdGenerator` with `PersistentIdGenerator` which uses a
simpler ID scheme and manages its own state. This allows the generator
to be reinitialized without first reading max ID information reported by
the queues.

This simplifies initialization of the ID generator and will allow the
queue initialization to be made significantly simpler. This also removes
the coupling between the ID generator from the queue data format.
Lastly, with `StandardIdGenerator` having been deleted there is no
longer a need to maintain the ID generator "epoch" as global state.