bugfix: Handle exhausted replacement segments
If `MergedSegmentReader.replaceSegments` is called with an immediately
exhausted segment (no values left to read) then the merged reader would
be left in an incorrect state in which the merged reader itself would
appear exhausted.

This was the result of not correctly restoring the "head" value in the
case of the replacement segment being immediately exhausted. An
additional call to `updateHead` has been added to handle this case.