Implement globally-aware segment flush strategy
Previously, in-memory writable segments were flushed when they had 128k
entries. This logic was always intended to be replaced as it suffered
from two major issues,

1. Entry size was not considered. The amount of memory for 128k entries
   varies significantly depending on the size of the entries.
2. The flush decision was made independently by each queue. With many
   busy queues we will need to flush each more often than if only one
   queue is busy. Independent queue flush decisions will inherently need
   to be conservative (i.e. err towards flushing often) to avoid memory
   exhaustion in the event there are other equally busy queues.

The new strategy implemented in this change is globally aware and based
on the size of entries. Global awareness is introduced through a new
`SegmentFlushScheduler` which receives updates from each queue about the
size of its current segment and returns an indication of whether the
queue should perform a flush.

The flush condition implemented by this change is relatively simple:
larger-than-average segments are flushed when the total size of
in-memory segments exceeds a configured threshold.