Add support for deleting queues
This required some fairly extensive refactoring to avoid leaking
references to already-deleted (or currently being deleted) queues.
Before this change, `QueueManager` would return references to `Queue`
objects which could then be used to execute queue operations (enqueue,
dequeue, etc). However, when deleting a queue it is important that we
first ensure the queue is idle which requires stronger ownership by the
`QueueManager` over queue lifecycle and access.

The solution implemented here is to refactor `QueueManager` such that a
queue-level lock is held while submiting operations to the queue or
deleting the queue. For non-delete operations it is only necessary to
hold the lock when starting an async operationg (e.g. we only hold the
lock while calling `enqueueAsync`, not while awaiting the corresponding
future). The delete process itself will wait for outstanding futures to
complete before proceeded (by virtue of the disruptor being shutdown
before the deletion commences).

The effect of this refactoring is that `QueueManager` no longer hands
out references to `Queue` objects, but instead provides methods for
applying user-provided asynchronous functions to the queue.

Another effect of this change is that all event submission to the
Disruptor ring buffer is now serialized by the queue-level locking. This
allows us to switch to the single-producer disruptor producer type, but
also suggests that the disruptor pattern is probably not a good fit for
Qrono.