qrono-bench: Reimplement as non-async
The async implementation added a lot of overhead, in part, due to
issuing each `enqueue` operation within its own async task. This new
implementation uses a simpler approach, relying on synchronous
operations, threading, and batching to obtain good throughput.

As a result, the new implementation is able to sustain much greater
throughput with lower CPU. This new implementation is able to easily
staturate a Qrono instance.