WIP: Rust implementation of Qrono
A work-in-progress implementation of Qrono in Rust now lives in the
qrono-rs directory. This implementation follows the same general design
and interface as the Java implementation, but makes various changes to
the implementation and on-disk formats.

Performance and tail latency of this new implementation is significantly
improved. The most significant missing features of the Rust
implementation currently are the lack of HTTP and gRPC interfaces, and
lack of working set compaction.