Change ID handling in enqueue and requeue
A new ID is generated when requeueing a value. This ensures requeue
order (instead of ID order) is respected when requeueing values with
equal deadlines. It also fixes a bug that occurs when a value is
requeued with its original deadline (as can occur when a value is
requeued within the same milliseconds).

IDs have also been removed from ENQUEUE responses. While including the
ID had tenuous value, it is made less useful by no longer being stable
across requeues.