Fix memory ordering with scheduling
It seems that the relaxed memory operations on the scheduler state can
be reordered with acquire/release operations against other variables.
This allows the logical sequence: "add item to queue, then schedule the
task" to be reordered as "schedule the task and add the item to the
queue".

The item in the queue can then end up not being processed until the next
schedule execution if an execution races with these operations (i.e.
"schedule task, task starts running, item is added to queue").