Fix various aliasing bugs in qrono-channel
Most of these were resolved by replacing internal &mut self receivers
with &self receivers (with documented safety rules) and keeping Blocks
as raw pointers until they are drained. Most of the corresponding
aliasing issues were due to the fact that Slots are written
asynchronously and the writes may not be complete when the Slot is
returned to the user in a Batch. Thus, we can not safely construct
mutable references to the Block or Slot until we are certain all writes
have completed.